Remove unit dependency on live tests
