--- tests/Clc.Polaris.Api.LiveIntegrationTests/Properties/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Clc.Polaris.Api.UnitTests")]
